Remove Init of Condition State and Factory
Remove :user_with_condition Factory as well as the initialization of the
Condition's state from the Condition model and change the Condition
migration to default the state to true. This is more likely to be the
scenario since when a user registers (currently), they are auto
signed-in, and they have their Condition created initially. This
eliminates duplicate Conditions from being created in the tests (and
potentially in the application itself).

NOTE: This commit requires a database reset as the migration change
alters the Condition model to set the 'state' attribute to 'true' by
default (as can be seen by the schema.rb changes).