Add Initial Setup for Mobile vs Non-Mobile
Add the layouts for the Mobile/Non-Mobile versions of the application,
to include mobile versions of some of the partials and views, to include
merging common code into global functions as a start (to keep it DRY).
Note that this is the first cut at the mobile app using jqMobile and can
be seen via navigating a mobile web browser to the server's location,
port 3000 (or setting the user-agent in a browser to use a mobile
platform).