Add Config to Prevent Initializing on Precompile
Add a config/application.rb directive that prevents Heroku from
initializing the application and attempting to connect to the database
during the asset precompile phase. Without this directive, Heroku will
inherently attempt to start the application and connect to the database,
but this will fail with a "Could not connect to localhost" error since
no environment variables are present during the precompile phase (which
indicate how to connect to the database).