Remove Faye from Project
Remove the Faye server and configuration from the project. This is
necessary because Heroku only allows 1 process per slug for the free
(non-paid) subscription, which means that a separate project needs to be
created to manage the Faye server portion of the application. The Faye
server project will be added as a submodule to the RCheckin base
project.