Add Faye Git Submodule from rcheckin-faye Repo
Add a submodule faye/ directory that is a clone of the rcheckin-faye git
project. This directory is what will be used for instantiating local
instances of the Faye server using Foreman when local development is to
occur. Note, however, that this is only for local development.
Production will have a separate Faye server running that is the
stand-alone rcheckin-faye project.