Add APP_CONFIG Configuration Parameter and File
Add an application_config.yml file that will keep track of
environment-specific variables that can be used depending on the
environment being run (i.e. production, test, development). This, for
now, is used for indicating the location of the Faye server URL required
to obtain the faye.js file (different between local instances of the
application and the Heroku deployed application).