Update Views/JS to Use APP_CONFIG Faye Parameter
Update all views, layouts, helpers and JavaScript code to utilize the
APP_CONFIG[:faye_url] configuration parameter to indicate where to get
the faye.js JavaScript file - this is because in production mode, the
URL will correspond to the Heroku-deployed Faye server (rcheckin-faye),
while local development and test instances will point to localhost. This
is the mechanism/standard that will be used moving forward for
environment-specific settings.